student and coach handler bugs fixed, jwt functions added

+++ models/entities/documents.go
@@ -7,7 +7,7 @@ type Document struct {
 	UploaderID uint
 	UniqueName string `gorm:"unique_name"` // hashli kısım
 	DocName    string `gorm:"size:64"`
-	Size       float32
+	Size       float64
 	Date       time.Time `gorm:"not null"`
 	DocType    string    `gorm:"doc_type"`
 	File       []byte    `gorm:"type:longblob"`
