+++ models/entities/meals.go
@@ -8,7 +8,7 @@ type Meal struct {
 	MealName  string 	`gorm:"size:32;not null"`
 	Description string 	`gorm:"size:128"` 
 	Kcal      uint		`gorm:"not null"`
-	Protein   float32
+	Protein   float64
 	Date      time.Time
-	Oil float32 
+	Oil float64 
 }
